Move non-component TypeScript files into separate directory
