+++ web/app/src/Checkbox.tsx
@@ -1,23 +1,6 @@
 import React, { useEffect, useRef } from 'react';
-import { mergeRefs } from './Utility';
-
-export enum CheckedState {
-  Unchecked,
-  Checked,
-  Indeterminate,
-}
-
-export namespace CheckedState {
-  export function fromBoolean(value: boolean) {
-    return value ? CheckedState.Checked : CheckedState.Unchecked;
-  }
-
-  /// Update a checkbox element according to a given checked state.
-  export function updateCheckbox(element: HTMLInputElement, checkedState: CheckedState) {
-    element.checked = checkedState === CheckedState.Checked;
-    element.indeterminate = checkedState === CheckedState.Indeterminate;
-  }
-}
+import { mergeRefs } from './common/Utility';
+import { CheckedState } from './common/CheckedState';
 
 export interface CheckboxProps {
   className?: string;
