+++ web/app/src/common/ConfigTypes.ts
