+++ web/app/src/common/JsonTypes.ts
