+++ web/app/src/common/Utility.ts
